+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
+## v4.X.X
+  - [EE] Fix the `:html` formatter: the default font size from a DOCX template is always applied on the rendered HTML.
+
+
 ## v4.15.4
   - [EE] Accept charts in PPTX without style/color description
   - [EE] Fix a minor PPTX generation issue when using chart